+++ pkg/cost-estimator/resource_types.go
@@ -1,6 +1,6 @@
 package cost_estimator
 
-type CostEstimatorFunc func(h *HttpHandler, resourceId string, timeInterval int64) (float64, error)
+type CostEstimatorFunc func(h *HttpHandler, resourceId string, timeInterval int) (float64, error)
 
 var azureResourceTypes = map[string]CostEstimatorFunc{
 	"Microsoft.Compute/virtualMachines": GetComputeVirtualMachineCost,
